push pyright CI config
