+++ .github/workflows/pyright.yml
@@ -1,0 +1,12 @@
+name: Pyright Check
+
+on: [push, pull_request]
+
+jobs:
+  pyright:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: jakebailey/pyright-action@v2
+        with:
+          version: 1.1.311 # Optional (change me!)
+          python-version: 3.11
