+++ .gitignore
@@ -1,3 +1,5 @@
 gpt-settings.talon
 # TalonDoc
 /docs/_build
+.ruff_cache/
+.mypy_cache/
