--- GPT/examples.md
@@ -1,9 +1,0 @@
-# Examples
-
-## Format Text Into a Markdown Table
-
-![Format raw text into a markdown table](./examples/table.gif)
-
-## Format Text Into a Bullet Point List
-
-![Format raw text into bullet points](./examples/bullets.gif)
