+++ GPT/examples/examples.md
@@ -1,0 +1,9 @@
+# Examples
+
+## Format Text Into a Markdown Table
+
+![Format raw text into a markdown table](./table.gif)
+
+## Format Text Into a Bullet Point List
+
+![Format raw text into bullet points](./bullets.gif)
