+++ docs/index.md
@@ -1,9 +1,13 @@
 # Documentation for talon-ai-tools
 
-This is a placeholder. Documentation is automatically generated using talondoc and Github Actions.
+This is a placeholder. Documentation is automatically generated using talondoc and Github Actions. See each folder for the associated README.md file which documents behavior, or the website at [https://colton.bio/talon-ai-tools/](https://colton.bio/talon-ai-tools/)
+
+## Generation
 
 To generate the documentation locally, run:
 
 ```sh
 python -X utf8 -m talondoc autogen -o docs --generate-index
 ```
+
+**NOTE**: To generate docs locally you need to move the conf.py and docs_root.md files to the root of the repo. This is normally done in CI to make the repo cleaner.
