+++ pyproject.toml
@@ -16,7 +16,7 @@ classifiers = [
 ]
 
 [tool.black]
-target-version = ['py310']
+target-version = ['py311']
 
 [tool.isort]
 profile = 'black'
@@ -25,4 +25,11 @@ profile = 'black'
 pythonpath = [".", "test/stubs"]
 
 [tool.pyright]
+# Talon classes don't use self so ignore the associated errors
 reportSelfClsParameterName = false
+reportGeneralTypeIssues = false
+# Imgui functions return an object that can't be statically resolved
+reportFunctionMemberAccess = false
+# Talon can't be installed in CI so ignore source errors
+reportMissingModuleSource = false
+reportMissingImports = false
